Refactor: Improve Sinoptico page and services
This commit refactors the Sinoptico page and its associated data services to improve modularity, readability, and maintainability.

Key changes include:

1.  **Service Splitting:** The monolithic `dataService.js` has been split into separate, domain-specific services for each module (`clientes`, `proveedores`, `insumos`, `productos`) located in `src/services/modules/`.

2.  **Custom Data Hook:** A new `useData` custom hook has been introduced to encapsulate data fetching logic, including loading and error state management. This simplifies the components that consume data.

3.  **Component Abstraction:** The AG Grid table has been extracted into a generic `DataGrid.jsx` component, reducing the complexity of `SinopticoPage.jsx`.

4.  **Code Modernization:** The `SinopticoPage.jsx` component has been updated to use the new hook and `DataGrid` component, making it cleaner and more focused on state management and view logic.
